third_party/WebKit/LayoutTests/media/encrypted-media/prefixed/encrypted-media-events-expected.txt

Test all the key-related events.

EXPECTED (messageEvent != 'null') OK
EXPECTED (messageEvent instanceof window.MediaKeyMessageEvent == 'true') OK
EXPECTED (keyNeededEvent != 'null') OK
EXPECTED (keyNeededEvent instanceof window.MediaKeyEvent == 'true') OK

*** Verify the presence of on* attributes. These would return undefined if they are not present. ***
EXPECTED (video.onwebkitkeyadded === 'null') OK
EXPECTED (video.onwebkitkeyerror === 'null') OK
EXPECTED (video.onwebkitkeymessage === 'null') OK
EXPECTED (video.onwebkitneedkey === 'null') OK

*** Test events using on* attributes. ***
needkey event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == '') OK
EXPECTED (event.sessionId == '') OK
EXPECTED (event.initData.length == '16') OK
EXPECTED (event.initData[0] == '48') OK
EXPECTED (event.initData[1] == '49') OK
EXPECTED (event.initData[2] == '50') OK
EXPECTED (event.initData[3] == '51') OK
EXPECTED (event.initData[4] == '52') OK
EXPECTED (event.initData[5] == '53') OK
EXPECTED (event.initData[6] == '54') OK
EXPECTED (event.initData[7] == '55') OK
EXPECTED (event.initData[8] == '56') OK
EXPECTED (event.initData[9] == '57') OK
EXPECTED (event.initData[10] == '48') OK
EXPECTED (event.initData[11] == '49') OK
EXPECTED (event.initData[12] == '50') OK
EXPECTED (event.initData[13] == '51') OK
EXPECTED (event.initData[14] == '52') OK
EXPECTED (event.initData[15] == '53') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode === 'null') OK
EXPECTED (event.systemCode == '0') OK

RUN(video.webkitGenerateKeyRequest('webkit-org.w3.clearkey', event.initData))
keymessage event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
The sessionId should be a non-empty string containing an integer.
EXPECTED (event.sessionId != '') OK
EXPECTED (event.sessionId != 'null') OK
EXPECTED (event.sessionId != 'undefined') OK
EXPECTED (isNaN(event.sessionId) == 'false') OK
EXPECTED (String(event.sessionId) == String(parseInt(event.sessionId)) == 'true') OK
Implementations should avoid sessionIds of 0.
EXPECTED (event.sessionId > '0') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message.length == '16') OK
EXPECTED (event.message[0] == '48') OK
EXPECTED (event.message[1] == '49') OK
EXPECTED (event.message[2] == '50') OK
EXPECTED (event.message[3] == '51') OK
EXPECTED (event.message[4] == '52') OK
EXPECTED (event.message[5] == '53') OK
EXPECTED (event.message[6] == '54') OK
EXPECTED (event.message[7] == '55') OK
EXPECTED (event.message[8] == '56') OK
EXPECTED (event.message[9] == '57') OK
EXPECTED (event.message[10] == '48') OK
EXPECTED (event.message[11] == '49') OK
EXPECTED (event.message[12] == '50') OK
EXPECTED (event.message[13] == '51') OK
EXPECTED (event.message[14] == '52') OK
EXPECTED (event.message[15] == '53') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode === 'null') OK
EXPECTED (event.systemCode == '0') OK

RUN(video.webkitAddKey('webkit-org.w3.clearkey', key, event.initData, event.sessionId))
keyadded event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
EXPECTED (event.sessionId == keyMessageSessionId == 'true') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode === 'null') OK
EXPECTED (event.systemCode == '0') OK

RUN(video.webkitAddKey('webkit-org.w3.clearkey', invalidKey, null, event.sessionId))
keyerror event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
EXPECTED (event.sessionId == keyMessageSessionId == 'true') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode.code == '1') OK
EXPECTED (event.systemCode == '0') OK

*** Test events using addEventListener(). ***
EVENT(webkitneedkey)
needkey event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == '') OK
EXPECTED (event.sessionId == '') OK
EXPECTED (event.initData.length == '16') OK
EXPECTED (event.initData[0] == '48') OK
EXPECTED (event.initData[1] == '49') OK
EXPECTED (event.initData[2] == '50') OK
EXPECTED (event.initData[3] == '51') OK
EXPECTED (event.initData[4] == '52') OK
EXPECTED (event.initData[5] == '53') OK
EXPECTED (event.initData[6] == '54') OK
EXPECTED (event.initData[7] == '55') OK
EXPECTED (event.initData[8] == '56') OK
EXPECTED (event.initData[9] == '57') OK
EXPECTED (event.initData[10] == '48') OK
EXPECTED (event.initData[11] == '49') OK
EXPECTED (event.initData[12] == '50') OK
EXPECTED (event.initData[13] == '51') OK
EXPECTED (event.initData[14] == '52') OK
EXPECTED (event.initData[15] == '53') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode === 'null') OK
EXPECTED (event.systemCode == '0') OK

RUN(video.webkitGenerateKeyRequest('webkit-org.w3.clearkey', event.initData))
EVENT(webkitkeymessage)
keymessage event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
The sessionId should be a non-empty string containing an integer.
EXPECTED (event.sessionId != '') OK
EXPECTED (event.sessionId != 'null') OK
EXPECTED (event.sessionId != 'undefined') OK
EXPECTED (isNaN(event.sessionId) == 'false') OK
EXPECTED (String(event.sessionId) == String(parseInt(event.sessionId)) == 'true') OK
Implementations should avoid sessionIds of 0.
EXPECTED (event.sessionId > '0') OK
The sessionsId should be different from the first run.
EXPECTED (event.sessionId != firstRunKeyMessageSessionId == 'true') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message.length == '16') OK
EXPECTED (event.message[0] == '48') OK
EXPECTED (event.message[1] == '49') OK
EXPECTED (event.message[2] == '50') OK
EXPECTED (event.message[3] == '51') OK
EXPECTED (event.message[4] == '52') OK
EXPECTED (event.message[5] == '53') OK
EXPECTED (event.message[6] == '54') OK
EXPECTED (event.message[7] == '55') OK
EXPECTED (event.message[8] == '56') OK
EXPECTED (event.message[9] == '57') OK
EXPECTED (event.message[10] == '48') OK
EXPECTED (event.message[11] == '49') OK
EXPECTED (event.message[12] == '50') OK
EXPECTED (event.message[13] == '51') OK
EXPECTED (event.message[14] == '52') OK
EXPECTED (event.message[15] == '53') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode === 'null') OK
EXPECTED (event.systemCode == '0') OK

RUN(video.webkitAddKey('webkit-org.w3.clearkey', key, event.initData, event.sessionId))
EVENT(webkitkeyadded)
keyadded event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
EXPECTED (event.sessionId == keyMessageSessionId == 'true') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode === 'null') OK
EXPECTED (event.systemCode == '0') OK

RUN(video.webkitAddKey('webkit-org.w3.clearkey', invalidKey, null, event.sessionId))
EVENT(webkitkeyerror)
keyerror event occurred
EXPECTED (event.target == '[object HTMLVideoElement]') OK
EXPECTED (event instanceof window.MediaKeyEvent == 'true') OK
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
EXPECTED (event.sessionId == keyMessageSessionId == 'true') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode.code == '1') OK
EXPECTED (event.systemCode == '0') OK

Attributes are read-only.
RUN(event.keySystem = 'blah')
RUN(event.sessionId = 'blah')
RUN(event.initData = new Uint8Array([0x12]))
RUN(event.message = new Uint8Array([0x12]))
RUN(event.defaultURL = 'example.com')
RUN(event.errorCode.code = MediaKeyError.MEDIA_KEYERR_CLIENT)
RUN(event.systemCode = 123)
EXPECTED (event.keySystem == 'webkit-org.w3.clearkey') OK
EXPECTED (event.sessionId == keyMessageSessionId == 'true') OK
EXPECTED (event.initData === 'null') OK
EXPECTED (event.message === 'null') OK
EXPECTED (event.defaultURL == '') OK
EXPECTED (event.errorCode.code == '1') OK
EXPECTED (event.systemCode == '0') OK
END OF TEST
